moved functions to separate file to enable unittest imports
